Add shell option for manual building

+++ framework/create-frameworks.sh
@@ -40,6 +40,7 @@ print_usage()
     echo "  -f        - Do not rebuild, just recreate the frameworks"
     echo "  -n        - Do not update modules (network-less mode)"
     echo "  -l        - Do not run 'make clean' before building"
+    echo "  -s        - Start a shell for manual building for each framework"
     echo "  -h        - Display this help text"
     echo "            - FRAMEWORK... is an optional list of frameworks to create"
     echo "              Valid framework names are: $all_modules"
@@ -57,6 +58,16 @@ uninstall_modules()
     done
 }
 
+install_modules()
+{
+    for m in $*; do
+        srcdir=`jhbuild gtk-osx-get-srcdir $m`
+        if [ "x$srcdir" != x ]; then
+            jhbuild run sh -c "cd $srcdir 2>/dev/null && make install"
+        fi
+    done
+}
+
 create_framework()
 {
     framework=$1
@@ -65,20 +76,27 @@ create_framework()
     if (echo "$modules" | grep -w $framework) >/dev/null; then
         modules=`echo $modules | sed -e 's/$framework//'`
 
+        if [ "x$shell" == xyes ]; then
+            jhbuild shell
+            exit 0
+        fi
+
         # Special-case WebKit, "make clean" breaks it...
         clean_save=$clean
         if [ "$*" == "WebKit" ]; then
             clean=
         fi
 
-        # Uninstall (from any previous attempts that failed) so the
-        # following modules don't link against the dylibs, but the
-        # frameworks instead.
-        uninstall_modules $*
-
         if [ $rebuild == yes ]; then
+            # Uninstall (from any previous attempts that failed) so
+            # the following modules don't link against the dylibs, but
+            # the frameworks instead.
+            uninstall_modules $*
+
             rm "$PREFIX"/lib/*.la 2>/dev/null
             jhbuild buildone $update $clean $* || exit 1
+        else
+            install_modules $*
         fi
 
         clean=$clean_save
@@ -89,14 +107,22 @@ create_framework()
         ./create-$framework-framework.sh $PREFIX || exit 1
 
         #cp -R $framework.framework $framework-runtime.framework || exit 1
+        # FIXME: This only removes the symlinks.
         #rm -rf $framework-runtime.framework/Headers
         #rm -rf $framework-runtime.framework/Resources/dev
         #strip ...
 
-        # Uninstall so the following modules don't link against the
-        # dylibs, but the frameworks instead.
-        uninstall_modules $*
+        if [ $rebuild == yes ]; then
+            # Uninstall so the following modules don't link against the
+            # dylibs, but the frameworks instead.
+            uninstall_modules $*
+        fi
     fi
+}
+
+use_framework()
+{
+    framework=$1
 
     if [ "x$JHB_PREPEND_FRAMEWORKS" == x ]; then
         export JHB_PREPEND_FRAMEWORKS=`pwd`/$framework.framework
@@ -105,10 +131,11 @@ create_framework()
     fi
 }
 
+shell=
 rebuild=yes
 update=
 clean=-c
-while getopts "fnlh" o; do
+while getopts "fnlsh" o; do
     case "$o" in
         f)
             rebuild=no
@@ -118,6 +145,9 @@ while getopts "fnlh" o; do
             ;;
         l)
             clean=
+            ;;
+        s)
+            shell=yes
             ;;
         h)
             print_usage
@@ -154,14 +184,22 @@ fi
 # We build gettext here instead of in bootstrap because it's part of
 # the GLib framework.
 create_framework GLib gettext-fw glib
+use_framework GLib
 
 create_framework Cairo pixman cairo
+use_framework Cairo
 
 # gnome-icon-theme requires gettext, that's why we build it here.
 create_framework Gtk gnome-icon-theme atk pango gtk+ gtk-engines ige-mac-integration
+use_framework Gtk
 
-#create_framework Libglade libglade
+exit 0
 
-#create_framework Loudmouth loudmouth
+create_framework Libglade libglade
+use_framework Libglade
 
-#create_framework WebKitGtk WebKit
+create_framework Loudmouth loudmouth
+use_framework Loudmouth
+
+create_framework WebKitGtk WebKit
+use_framework WebKitGtk
